pyproject.toml
[project]
name = "travel-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.40.0",
    "streamlit>=1.56.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
]
